+++ perf/tests/test_runner.py
@@ -65,7 +65,7 @@ class TestRunner(unittest.TestCase):
     def test_worker(self):
         result = self.exec_runner('--worker')
         self.assertRegex(result.stdout,
-                         r'^bench: Median \+- MAD: 1\.00 sec \+- 0\.00 sec\n$')
+                         r'^bench: Mean \+- std dev: 1\.00 sec \+- 0\.00 sec\n$')
 
     def test_debug_single_value(self):
         result = self.exec_runner('--debug-single-value', '--worker')
@@ -116,7 +116,7 @@ class TestRunner(unittest.TestCase):
                          r'Metadata:\n'
                          r'(?:- .*\n)+'
                          r'\n'
-                         r'bench: Median \+- MAD: 1\.00 sec \+- 0\.00 sec\n$')
+                         r'bench: Mean \+- std dev: 1\.00 sec \+- 0\.00 sec\n$')
 
     def test_loops_calibration(self):
         def time_func(loops):
@@ -298,11 +298,11 @@ class TestRunner(unittest.TestCase):
     def test_show_name(self):
         result = self.exec_runner('--worker', name='NAME')
         self.assertRegex(result.stdout,
-                         r'^NAME: Median \+- MAD: 1\.00 sec \+- 0\.00 sec\n$')
+                         r'^NAME: Mean \+- std dev: 1\.00 sec \+- 0\.00 sec\n$')
 
         result = self.exec_runner('--worker', name='NAME', show_name=False)
         self.assertRegex(result.stdout,
-                         r'^Median \+- MAD: 1\.00 sec \+- 0\.00 sec\n$')
+                         r'^Mean \+- std dev: 1\.00 sec \+- 0\.00 sec\n$')
 
     def test_compare_to(self):
         def time_func(loops):